source/requirements.txt
numpy
tensorflow-gpu==1.13.2
tflearn==0.3.2
opencv-python
pyyaml
scipy
matplotlib
#Dependencies for goICP
Cython
autowrap
py-goicp
scikit-learn
trimesh